split code into parts
